+++ app/nodecron.js
@@ -3,7 +3,11 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const moment = require("moment");
 var shell = require("shelljs");
 var cron = require('node-cron');
+let firstProcessPid = null;
+let pm2Name = '???';
 async function cronjob(fastify) {
+    firstProcessPid = await firstPM2InstancePID();
+    console.log(`First process ID of '${pm2Name}' is ${firstProcessPid}`);
     const secondNow = +moment().get('second');
     const timingSch = `${secondNow} */1 * * * *`;
     let timingSchedule = [];
@@ -15,7 +19,7 @@ async function cronjob(fastify) {
     timingSchedule['isonline'].hour = process.env.IS_AUTO_SEND_EVERY_HOUR ? parseInt(process.env.IS_AUTO_SEND_EVERY_HOUR) : 0;
     timingSchedule['isonline'].hour = timingSchedule['isonline'].hour > 23 ? (timingSchedule['isonline'].hour % 23) : timingSchedule['isonline'].hour;
     timingSchedule['isonline'].minute = timingSchedule['isonline'].minute + timingSchedule['isonline'].hour * 60;
-    timingSchedule['isonline'].minute = timingSchedule['isonline'].minute < 20 ? 20 : timingSchedule['isonline'].minute;
+    timingSchedule['isonline'].minute = timingSchedule['isonline'].minute < 10 ? 10 : timingSchedule['isonline'].minute;
     if (timingSchedule['isonline'].minute <= 0) {
         timingSchedule['isonline'].autosend = false;
     }
@@ -23,7 +27,7 @@ async function cronjob(fastify) {
     timingSchedule['nrefer'].minute = process.env.NREFER_AUTO_SEND_EVERY_MINUTE ? parseInt(process.env.NREFER_AUTO_SEND_EVERY_MINUTE) : 0;
     timingSchedule['nrefer'].hour = process.env.NREFER_AUTO_SEND_EVERY_HOUR ? parseInt(process.env.NREFER_AUTO_SEND_EVERY_HOUR) : 0;
     timingSchedule['nrefer'].minute = timingSchedule['nrefer'].minute + timingSchedule['nrefer'].hour * 60;
-    timingSchedule['nrefer'].minute = timingSchedule['nrefer'].minute < 5 ? 5 : timingSchedule['nrefer'].minute;
+    timingSchedule['nrefer'].minute = timingSchedule['nrefer'].minute < 3 ? 3 : timingSchedule['nrefer'].minute;
     if (timingSchedule['nrefer'].minute <= 0) {
         timingSchedule['nrefer'].autosend = false;
     }
@@ -33,18 +37,21 @@ async function cronjob(fastify) {
             (process.env.HIS_DATACENTER_SEND_EVERY_HOUR ? +process.env.HIS_DATACENTER_SEND_EVERY_HOUR : 2) * 60;
     timingSchedule['cupDataCenter'].minute = timingSchedule['cupDataCenter'].minute < 20 ? 20 : timingSchedule['cupDataCenter'].minute;
     const minuteSinceLastNight = (+moment().get('hour')) * 60 + (+moment().get('minute'));
-    console.log(moment().format('HH:mm:ss'), " Start API for Hospcode", process.env.HOSPCODE);
-    console.log('crontab start: ', timingSch, 'minuteSinceLastNight', minuteSinceLastNight);
-    if (timingSchedule['nrefer'].autosend) {
+    if (firstProcessPid === process.pid) {
+        console.log(moment().format('HH:mm:ss'), " Start API for Hospcode", process.env.HOSPCODE);
+        console.log('crontab start: ', timingSch, 'minuteSinceLastNight', minuteSinceLastNight, `on process ID ${process.pid} of '${pm2Name}'`);
+    }
+    if (firstProcessPid === process.pid && timingSchedule['nrefer'].autosend) {
         console.log('crontab nRefer start every', timingSchedule['nrefer'].minute, ' (minute) from midnight.');
     }
-    if (timingSchedule['isonline'].autosend) {
+    if (firstProcessPid === process.pid && timingSchedule['isonline'].autosend) {
         console.log('crontab ISOnline start every', timingSchedule['isonline'].minute, ' (minute) from midnight.');
     }
-    if (timingSchedule['cupDataCenter'].autosend) {
+    if (firstProcessPid === process.pid && timingSchedule['cupDataCenter'].autosend) {
         console.log('crontab Data Center start every', timingSchedule['cupDataCenter'].minute, ' (minute) from midnight.');
     }
     cron.schedule(timingSch, async (req, res) => {
+        firstProcessPid = await firstPM2InstancePID();
         const minuteSinceLastNight = (+moment().get('hour')) * 60 + (+moment().get('minute'));
         const minuteNow = +moment().get('minute') == 0 ? 60 : +moment().get('minute');
         const hourNow = +moment().get('hour');
@@ -65,16 +72,6 @@ async function cronjob(fastify) {
         }
     });
     async function doAutoSend(req, res, serviceName, functionName) {
-        let firstProcessPid = { pid: -1 };
-        if (process.env.START_TOOL === 'nodemon') {
-            firstProcessPid = process.pid;
-        }
-        else {
-            if (!global.firstProcessPid) {
-                await getFirstProcessPid();
-            }
-            firstProcessPid = global.firstProcessPid ? global.firstProcessPid : -1;
-        }
         if (firstProcessPid === process.pid) {
             const now = moment().locale('th').format('HH:mm:ss');
             const db = serviceName == 'isonline' ? global.dbISOnline : global.dbHIS;
@@ -82,21 +79,26 @@ async function cronjob(fastify) {
             await require(functionName)(req, res, db, timingSchedule[serviceName]);
         }
     }
-    async function getFirstProcessPid() {
-        var jlist = await shell.exec('pm2 jlist', { silent: true });
-        let pm2Process = jlist && jlist !== '' ? JSON.parse(jlist) : [];
-        let processList = [];
-        for (let p of pm2Process) {
-            if (p.name == process.env.PM2_NAME) {
-                processList.push(p);
-            }
-        }
-        if (processList.length) {
-            global.firstProcessPid = processList[0].pid;
-        }
-    }
     async function getmophUrl() {
         global.mophService = await require('./routes/main/crontab')(global.mophService, {});
+    }
+    async function firstPM2InstancePID() {
+        var jlist = await shell.exec('pm2 jlist', { silent: true });
+        let pm2Process = jlist && jlist !== '' && jlist.length > 32 ? JSON.parse(jlist) : [];
+        pm2Name = '???';
+        if (pm2Process && pm2Process.length) {
+            const prc = pm2Process.filter((o) => process.pid == o.pid && o.pm2_env.status == 'online');
+            pm2Name = prc && prc.length ? prc[0].name : '???';
+        }
+        for (let procss of pm2Process) {
+            if (procss.name == pm2Name) {
+                firstProcessPid = firstProcessPid ? firstProcessPid : procss.pid;
+            }
+        }
+        if (firstProcessPid == process.pid) {
+            global.firstProcessPid = firstProcessPid;
+        }
+        return firstProcessPid;
     }
 }
 exports.default = cronjob;